requirements.txt
annotated-types==0.6.0
appdirs==1.4.4
arabic_reshaper==3.0.0
argcomplete==2.0.0
astroid==3.1.0
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
autocommand==2.2.2
backcall==0.2.0
beautifulsoup4==4.12.3
blosc2==2.5.1
btrfsutil==6.7.1
CacheControl==0.14.0
certifi==2022.12.7
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.3
colorama==0.4.6
contourpy==1.2.0
coverage==7.4.1
cryptography==42.0.5
cssselect==1.2.0
cycler==0.12.1
Cython==3.0.9
daemonize==2.5.0
dbus-python==1.3.2
decorator==4.4.2
dill==0.3.8
distro==1.9.0
dnspython==2.5.0
docopt==0.6.2
dukpy==0.3
elementpath==4.4.0
esprima==4.0.1
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.0
fastjsonschema==2.19.1
ffmpeg-python==0.2.0
filelock==3.12.0
fonttools==4.50.0
freetype-py==2.4.0
fsspec==2023.4.0
ftfy==6.1.1
future==0.18.3
gevent==23.9.1
gitdb==4.0.11
GitPython==3.1.42
glfw==2.6.5
greenlet==3.0.3
html5lib==1.1
httplib2==0.22.0
huggingface-hub==0.14.1
idna==3.4
imageio==2.34.0
imageio-ffmpeg==0.4.9
importlib_metadata==5.1.0
inflect==7.0.0
iniconfig==2.0.0
ipython==8.22.2
isort==5.13.2
jaraco.classes==3.3.1
jaraco.context==4.3.0
jaraco.functools==4.0.0
jaraco.text==3.12.0
javascripthon==0.12
jedi==0.19.1
jeepney==0.8.0
Jinja2==3.1.2
json_tricks==3.15.5
keyring==24.3.0
keyutils==0.6
kiwisolver==1.4.5
lensfun==0.3.4
lit==17.0.6.dev0
lockfile==0.12.2
lxml==5.1.0
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.8.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
meson==1.4.0
more-itertools==10.2.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.5
msgpack-numpy==0.4.8
ndindex==1.7
networkx==3.1
notify2==0.3.1
numexpr==2.9.0
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
open-clip-torch @ git+https://github.com/mlfoundations/open_clip.git@bb6e834e9c70d9c27d0dc3ecedeebeaeb1ffad6b
openpyxl==3.1.2
openrazer==3.7.0
openrazer-daemon==3.7.0
ordered-set==4.1.0
packaging==23.2
pandas==1.5.3
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==4.2.0
pluggy==1.4.0
ply==3.11
pooch==1.8.1
powerline-status==2.8.3
proglog==0.1.10
prompt-toolkit==3.0.43
protobuf==3.20.0
psutil==5.9.8
PsychoPy==2021.2.3
psychtoolbox==3.0.18.2
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyaml==23.12.0
pycairo==1.26.0
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
pydub==0.25.1
pyfakefs==5.3.5
pygame==2.5.2
pyglet==2.0.10
Pygments==2.17.2
PyGObject==3.46.0
pylint==3.1.0
pyo==1.0.5
PyOpenGL==3.1.6
pyOpenSSL==24.0.0
pyosf==1.0.5
pyparallel==0.2.2
pyparsing==3.1.2
PyQt5==5.15.10
PyQt5-sip==12.13.0
pyserial==3.5
pytest==7.4.4
python-bidi==0.4.2
python-dateutil==2.9.0
python-gitlab==4.4.0
python-vlc==3.0.20123
python-xlib==0.33
pytz==2024.1
pyudev==0.24.1
pyxdg==0.28
PyYAML==6.0.1
pyzmq==25.1.2
questplus==2019.4
ranger-fm==1.9.3
regex==2023.5.5
requests==2.30.0
requests-toolbelt==1.0.0
rich==13.3.5
scipy==1.12.0
scour==0.38.2
SecretStorage==3.3.3
sentencepiece==0.1.99
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sounddevice==0.4.6
soundfile==0.12.1
soupsieve==2.5
stack-data==0.6.3
sympy==1.11.1
tables==3.9.2
TBB==0.2
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.4
tqdm==4.66.2
traitlets==5.14.2
transformers==4.28.1
triton==2.0.0
trove-classifiers==2024.3.6
typing_extensions==4.10.0
udiskie==2.5.2
ufw==0.36.2
ujson==5.9.0
urllib3==2.0.2
validate-pyproject==0.13.post1.dev0+gb752273.d20230520
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.7.0
whisper-openai==1.0.0
wxPython==4.2.1
xarray==2023.1.0
xlrd==2.0.1
xmlschema==2.5.1
zipp==3.17.0
zope.event==5.0
zope.interface==6.2
zstandard==0.22.0
